created exporter for billing spreadsheet
