+++ src/main/java/work/billing/Export/BillingExporterThread.java
@@ -1,0 +1,32 @@
+package work.billing.Export;
+
+import work.billing.Cache.BaseSpreadSheetMatrix;
+import work.billing.Spreadsheets.COL;
+import work.billing.Spreadsheets.Spreadsheet;
+
+import java.util.AbstractMap;
+
+/**
+ * Created by mhillbrand on 3/6/2016.
+ */
+public class BillingExporterThread implements Runnable {
+    private final String workSheetName;
+    private final Spreadsheet spreadsheet;
+    private final BaseSpreadSheetMatrix cachedMatrix;
+
+    public BillingExporterThread(BaseSpreadSheetMatrix cachedMatrix, Spreadsheet spreadsheet, String worksheetName) {
+        this.cachedMatrix = cachedMatrix;
+        this.spreadsheet = spreadsheet;
+        this.workSheetName = worksheetName;
+    }
+
+    @Override
+    public void run() {
+        for (COL col : cachedMatrix.cellMatrix.keySet()) {
+            AbstractMap<Integer, String> rows = cachedMatrix.cellMatrix.get(col);
+            for (Integer row : rows.keySet()) {
+                spreadsheet.insertValueIntoCell(this.workSheetName, col.ordinal() , row, rows.get(row));
+            }
+        }
+    }
+}
